// src/app/common/models/vehiculo.ts
export class Vehiculo{
    idVEHICULO: number;
    placa: string;
    idCATEGORIAVEHICULO: number;
    marca: string;
    modelo: string;
    color: string;
    motor: string;
    vin: string;
    serie: string;
    anioFabricacion: number;
    anioModelo: number;
    carroceria: string;
    propietario: string;
    nroTarjeta: string;
    idTIPOVEHICULO: number;
    idTIPOSERVICIOVEHICULO: number;
    estado: boolean;

}